refator AllItem.tsx
